[ENTESB-1031] Improved patch functionality - offline patching - patch survival of data direction deletion - purge of old artefacts when patching offline
