+++ fabric/fabric-core/src/test/java/io/fabric8/service/PatchServiceImplTest.java
@@ -17,13 +17,8 @@
 package io.fabric8.service;
 
 import java.io.IOException;
-import java.util.Arrays;
-import java.util.List;
 import java.util.Properties;
-import java.util.Set;
 
-import org.apache.felix.utils.version.VersionTable;
-import io.fabric8.api.Patch;
 import org.junit.Ignore;
 import org.junit.Test;
 
@@ -32,49 +27,6 @@ import static junit.framework.Assert.assertTrue;
 
 @Ignore("[FABRIC-529] Fix fabric core PatchServiceImplTest")
 public class PatchServiceImplTest {
-
-    @Test
-    public void testVersionComparison() {
-        assertSmaller("2.5.0.fuse-beta-7-039", "2.5.0.fuse-70-070");
-        assertSmaller("2.5.0.fuse-beta-7-039", "2.5.0.fuse-beta-7-042");
-        assertSmaller("2.5.0.fuse-beta-70-039", "2.5.0.fuse-beta-71-010");
-        assertSmaller("2.5.0.fuse-7-061", "2.5.0.fuse-70-070");
-        assertSmaller("2.5.0.fuse-7-061", "2.5.0.fuse-7-0-068");
-        assertSmaller("2.5.0.fuse-7-0-061", "2.5.0.fuse-70-068");
-        assertSmaller("2.5.0.fuse-7-0-061", "2.5.0.fuse-71-018");
-    }
-
-    private void assertSmaller(String o1, String o2) {
-        assertTrue(o1 + " < " + o2, compare(o1, o2) < 0);
-    }
-
-    private int compare(String o1, String o2) {
-        org.osgi.framework.Version v1 = VersionTable.getVersion(o1);
-        org.osgi.framework.Version v2 = VersionTable.getVersion(o2);
-        return PatchServiceImpl.compareFuseVersions(v1, v2);
-    }
-
-    @Test
-    public void testDownload() throws Exception {
-        System.setProperty("karaf.home", "target/home");
-        System.setProperty("karaf.default.repository", "system");
-        System.setProperty("fuse.patch.location", "target/patches");
-
-        PatchServiceImpl service = new PatchServiceImpl(null, null, null, null);
-
-        List<String> repos = Arrays.asList("https://repo.fusesource.com/nexus/content/repositories/ea");
-
-        long t0 = System.currentTimeMillis();
-        Set<Patch> patches1 = service.loadPerfectusPatches(repos, true);
-        long t1 = System.currentTimeMillis();
-        Set<Patch> patches2 = service.loadPerfectusPatches(repos, false);
-        long t2 = System.currentTimeMillis();
-
-        assertEquals(patches1.size(), patches2.size());
-        assertTrue(t2 - t1 < (t1 - t0) / 2);
-
-        System.out.println(patches1);
-    }
 
     @Test
     public void testPatchDescriptorWithoutDirectives() throws IOException {
